requirements.txt
appnope==0.1.0 ; platform_system == 'Darwin'
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cvxopt==1.2.4
cycler==0.10.0
cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
dlx==1.0.4
docloud==1.0.375
docplex==2.12.182
entrypoints==0.3
fastdtw==0.3.4
future==0.18.2
grakel-dev==0.1a6
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0 ; python_version < '3.8'
inflection==0.3.1
ipykernel==5.1.4
ipython-genutils==0.2.0
ipython==7.12.0 ; python_version >= '3.3'
ipywidgets==7.5.1
jedi==0.16.0
jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyter==1.0.0
jupytext==1.3.3
kiwisolver==1.1.0
llvmlite==0.31.0
markupsafe==1.1.1
marshmallow-polyfield==5.8
marshmallow==3.4.0
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.2.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.2.3
networkx==2.4 ; python_version > '3.5'
nose==1.3.7
notebook==6.0.3
ntlm-auth==1.4.0
numba==0.48.0
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.6.7
ptyprocess==0.6.0 ; os_name != 'nt'
py==1.8.1
pycparser==2.19
pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pyscf==1.7.0 ; sys_platform != 'win32'
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3
pyzmq==18.1.1
qiskit-aer==0.3.4
qiskit-aqua==0.6.2
qiskit-ibmq-provider==0.4.5
qiskit-ignis==0.2.0
qiskit-terra==0.11.1
qiskit==0.14.1
qtconsole==4.6.0
quandl==3.5.0
requests-ntlm==1.1.0
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
send2trash==1.5.0
six==1.14.0
sklearn==0.0
sympy==1.5.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
umap-learn==0.3.10
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
websockets==7.0
wheel==0.34.2
widgetsnbextension==3.5.1
zipp==2.1.0
